+++ src/tests/RzR.DataVigil.EFCore.Tests/Entities/AuditOutboxRecord.cs
@@ -1,0 +1,11 @@
+namespace RzR.DataVigil.EFCore.Tests.Entities
+{
+    public class AuditOutboxRecord
+    {
+        public int Id { get; set; }
+
+        public string Topic { get; set; }
+
+        public string Payload { get; set; }
+    }
+}